# src/Tools/Refitter.Generator/OpenApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trend.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/News/Deactivate/{articleId}": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "Deactivate",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/News/Activate/{articleId}": {
      "get": {
        "tags": [
          "News"
        ],
        "operationId": "Activate",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/SearchWord/GetActiveSearchWords": {
      "get": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "GetActiveSearchWords",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/GetDeactivatedSearchWords": {
      "get": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "GetDeactivatedSearchWords",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<SearchWordResDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchWordResDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/AddNew": {
      "post": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "AddNew",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordAddReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordAddReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordAddReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/Filter": {
      "post": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordFilterReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordFilterReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchWordFilterReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchWordResDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/AttachImage/{searchWordId}": {
      "post": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "AttachImage",
        "parameters": [
          {
            "name": "searchWordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/Deactivate/{id}": {
      "delete": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "Deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/SearchWord/Activate/{id}": {
      "get": {
        "tags": [
          "SearchWord"
        ],
        "operationId": "Activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SearchWordAddReqDto": {
        "type": "object",
        "properties": {
          "searchWord": {
            "type": "string",
            "nullable": true
          },
          "searchEngine": {
            "type": "integer",
            "format": "int32"
          },
          "contextType": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SearchWordFilterReqDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "take": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "integer",
            "format": "int32"
          },
          "contextType": {
            "type": "integer",
            "format": "int32"
          },
          "searchEngine": {
            "type": "integer",
            "format": "int32"
          },
          "query": {
            "type": "string",
            "nullable": true
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SearchWordResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "searchWord": {
            "type": "string",
            "nullable": true
          },
          "searchEngineName": {
            "type": "string",
            "nullable": true
          },
          "searchEngineId": {
            "type": "integer",
            "format": "int32"
          },
          "contextTypeName": {
            "type": "string",
            "nullable": true
          },
          "contextTypeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:8080/realms/PortfolioRealm/protocol/openid-connect/auth",
            "scopes": { }
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}