# src/Web/admin-dashboard/apps/trend-dashboard/src/app/shared/services/open-api/model/search-word-add-req-dto.ts
/**
 * Trend.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SearchWordAddReqDto { 
    searchWord?: string | null;
    searchEngine?: number;
    contextType?: number;
}