# src/Web/admin-dashboard/apps/trend-dashboard/src/app/shared/services/open-api/model/search-word-res-dto.ts
/**
 * Trend.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SearchWordResDto { 
    id?: string | null;
    isActive?: boolean;
    created?: string;
    searchWord?: string | null;
    searchEngineName?: string | null;
    searchEngineId?: number;
    contextTypeName?: string | null;
    contextTypeId?: number;
}